Explicitly declare BOM dependencies in androidTestImplementation
Since AGP 4.1, handling of dependency graphs has changed a bit. That means that declaring version constraints through e.g. a BOM in the main dependency graph (`implementation`) doesn't apply them to the `androidTestImplementation` dependency graph, so we need to explicitly depend on the BOM.

We should revert this as soon as we can as this also means that any libraries depending on us through a BOM will have to declare a dependency on it in their android test graph if they have instrumented tests with Firebase dependencies.

See b/182167029 for an issue to track.